Live Report: create a component for Home
This will be required when we implement the router
Also, move FindingDetails and Utils to separate folders
